third_party/blink/web_tests/external/wpt/mediacapture-depth/dictionary-manual.https-expected.txt

This is a testharness.js-based test.
FAIL MediaTrackSupportedConstraints dictionary include attributes are correct assert_equals: Expect default value of dictionary.focalLengthX to be false expected false but got true
FAIL MediaTrackCapabilities dictionary of depth include attributes are correct assert_equals: Expect dictionary.videoKind to be string expected "string" but got "undefined"
FAIL MediaTrackConstraintSet dictionary of depth include attributes are correct assert_equals: Expect dictionary.focalLengthX to be number expected "number" but got "undefined"
FAIL MediaTrackSettings dictionary of depth include attributes are correct assert_equals: Expect dictionary.focalLengthX to be number expected "number" but got "undefined"
FAIL MediaTrackCapabilities dictionary of color include attributes are correct assert_equals: Expect dictionary.videoKind to be string expected "string" but got "undefined"
FAIL MediaTrackConstraintSet dictionary of color include attributes are correct assert_equals: Expect dictionary.focalLengthX to be number expected "number" but got "undefined"
FAIL MediaTrackSettings dictionary of color include attributes are correct assert_equals: Expect dictionary.focalLengthX to be number expected "number" but got "undefined"
Harness: the test ran to completion.
